패키지명 변경 및 firebase 패키지명 변경부분 적용
