+++ app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:464765094243:android:9d348f652fe8356e1d6b56",
         "android_client_info": {
-          "package_name": "com.example.random_chatting"
+          "package_name": "com.kj.random_chatting"
         }
       },
       "oauth_client": [
