+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.random_chatting">
+    package="com.kj.random_chatting">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.Randomchatting"
         tools:targetApi="31">
         <activity
-            android:name=".common.SplashActivity"
+            android:name="com.kj.random_chatting.common.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.NoActionBar">
 
@@ -30,12 +30,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".userRegist.MainActivity"
+            android:name="com.kj.random_chatting.userRegist.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
         </activity>
-        <activity android:name=".userRegist.FileUploadActivity" />
-        <activity android:name=".userList.UserListActivity" />
+        <activity android:name="com.kj.random_chatting.userRegist.FileUploadActivity" />
+        <activity android:name="com.kj.random_chatting.userList.UserListActivity" />
 
     </application>
