--- app/src/main/java/com/example/random_chatting/common/SplashActivity.java
@@ -1,28 +1,0 @@
-package com.example.random_chatting.common;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-import com.example.random_chatting.R;
-import com.example.random_chatting.userRegist.MainActivity;
-
-public class SplashActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash);
-
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent = new Intent(SplashActivity.this, MainActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        },2000);
-    }
-}
