--- app/src/main/java/com/example/random_chatting/ui/home/HomeViewModel.java
@@ -1,18 +1,0 @@
-package com.example.random_chatting.ui.home;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class HomeViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public HomeViewModel() {
-        mText = new MutableLiveData<>();
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
