+++ app/src/main/java/com/kj/random_chatting/util/ImageSliderAdapter.java
@@ -1,0 +1,57 @@
+package com.kj.random_chatting.util;
+
+import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.bumptech.glide.Glide;
+import com.kj.random_chatting.R;
+
+public class ImageSliderAdapter extends RecyclerView.Adapter<ImageSliderAdapter.MyViewHolder> {
+    private Context context;
+    private String[] sliderImage;
+
+    public ImageSliderAdapter(Context context, String[] sliderImage) {
+        this.context = context;
+        this.sliderImage = sliderImage;
+    }
+
+    @NonNull
+    @Override
+    public MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(parent.getContext())
+                .inflate(R.layout.item_slider, parent, false);
+        return new MyViewHolder(view);
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull MyViewHolder holder, int position) {
+        holder.bindSliderImage(sliderImage[position]);
+    }
+
+    @Override
+    public int getItemCount() {
+        return sliderImage.length;
+    }
+
+    public class MyViewHolder extends RecyclerView.ViewHolder {
+
+        private ImageView mImageView;
+
+        public MyViewHolder(@NonNull View itemView) {
+            super(itemView);
+            mImageView = itemView.findViewById(R.id.imageSlider);
+        }
+
+        public void bindSliderImage(String imageURL) {
+            Glide.with(context)
+                    .load(imageURL)
+                    .into(mImageView);
+        }
+    }
+}
