+++ app/src/main/java/com/kj/random_chatting/util/Retrofit_client.java
@@ -1,0 +1,23 @@
+package com.kj.random_chatting.util;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+import retrofit2.converter.scalars.ScalarsConverterFactory;
+
+public class Retrofit_client {
+    private static final String BASE_URL = "https://random-chatting-server.herokuapp.com/";
+
+    public static Retrofit_interface getApiService(){return getInstance().create(Retrofit_interface.class);}
+
+    private static Retrofit getInstance(){
+        Gson gson = new GsonBuilder().setLenient().create();
+        return new Retrofit.Builder()
+                .baseUrl(BASE_URL)
+                .addConverterFactory(ScalarsConverterFactory.create())
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .build();
+    }
+}
