+++ app/src/main/java/com/kj/random_chatting/util/ViewPagerClass.java
@@ -1,0 +1,57 @@
+package com.kj.random_chatting.util;
+
+import android.content.Context;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
+
+import androidx.core.content.ContextCompat;
+
+import com.kj.random_chatting.R;
+
+public class ViewPagerClass {
+    Context viewPagerContext;
+    LinearLayout layoutIndicator;
+
+    public ViewPagerClass (Context context, LinearLayout layout) {
+        viewPagerContext = context;
+        layoutIndicator = layout;
+    }
+
+    //slider
+    public void setupIndicators(int count) {
+        layoutIndicator.removeAllViews();
+        ImageView[] indicators = new ImageView[count];
+        LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(
+                ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
+
+        params.setMargins(16, 8, 16, 8);
+
+        for (int i = 0; i < indicators.length; i++) {
+            indicators[i] = new ImageView(viewPagerContext);
+            indicators[i].setImageDrawable(ContextCompat.getDrawable(viewPagerContext,
+                    R.drawable.bg_indicator_inactive));
+            indicators[i].setLayoutParams(params);
+            layoutIndicator.addView(indicators[i]);
+        }
+        setCurrentIndicator(0);
+    }
+
+    public void setCurrentIndicator(int position) {
+        int childCount = layoutIndicator.getChildCount();
+        for (int i = 0; i < childCount; i++) {
+            ImageView imageView = (ImageView) layoutIndicator.getChildAt(i);
+            if (i == position) {
+                imageView.setImageDrawable(ContextCompat.getDrawable(
+                        viewPagerContext,
+                        R.drawable.bg_indicator_active
+                ));
+            } else {
+                imageView.setImageDrawable(ContextCompat.getDrawable(
+                        viewPagerContext,
+                        R.drawable.bg_indicator_inactive
+                ));
+            }
+        }
+    }
+}
