+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ic_launcher_background"
-    tools:context=".common.SplashActivity">
+    tools:context="com.kj.random_chatting.common.SplashActivity">
 
     <ImageView
         android:id="@+id/imageView"
