+++ app/src/main/res/layout/user_list_activity.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".userRegist.MainActivity">
+    tools:context="com.kj.random_chatting.userRegist.MainActivity">
 
 
     <androidx.viewpager2.widget.ViewPager2
